[opt] 结构优化

+++ src/main/java/com/tianli/charge/vo/OrderFundTransactionVO.java
@@ -28,7 +28,7 @@ public class OrderFundTransactionVO extends OrderBaseVO {
     /**
      * 年化率
      */
-    private BigDecimal rate;
+    private double rate;
 
     /**
      * id
